pyproject.toml
[project]
name = "mn-ai-voice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "alembic>=1.18.1",
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pytest>=9.0.2",
    "python-ulid>=3.1.0",
    "pyyaml>=6.0.3",
    "sqlalchemy>=2.0.45",
    "sqlalchemy-orm>=1.2.10",
    "types-pyyaml>=6.0.12.20250915",
    "uvicorn>=0.40.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=9.0.2",
  "httpx>=0.28.1",
]

[tool.pytest.ini_options]
pythonpath = ["."]

[tool.mypy]
python_version = "3.12"
install_types = true
non_interactive = true

[[tool.mypy.overrides]]
module = [
    "yaml.*",
    "yaml"
]
follow_imports = "silent"

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "types-pyyaml>=6.0.12.20250915",
]
